split allocatins in two diffrent sets
